+++ src/main/java/duke/utils/TaskList.java
@@ -1,4 +1,7 @@
-import TaskTypes.Task;
+package duke.utils;
+
+import duke.DukeException;
+import duke.task.Task;
 
 import java.util.ArrayList;
