+++ text-ui-test/EXPECTED.TXT
@@ -27,5 +27,5 @@ Your tasks:
 1. [T] [ ] make bed
 2. [E] [ ] concert (at: tomorrow)
 3. [D] [ ] cs2103t submission (by: friday 2pm)
-DukeException: You did not provide a valid command
+duke.DukeException: You did not provide a valid command
 Goodbye. Hope to see you again soon!
